+++ src/Hedwig/ClientApp/src/utils/modelUtils/index.ts
@@ -1,0 +1,2 @@
+export * from './genderUtils';
+export * from './childUtils';
